backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
ecdsa==0.19.0 
fastapi==0.115.5
fastapi-utils==0.8.0
frozenlist==1.5.0
gotrue==2.10.0
greenlet==3.1.1
grpcio==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.0.1
idna==3.10
jiter==0.7.1
python-jose==3.3.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab==4.3.1
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-community==2.0.2
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.143
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.54.5
orjson==3.10.11
packaging==24.2
postgrest==0.18.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
realtime==2.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.2
